+++ Examples/source/Sequence/TMU/Code Modules/TestSteps/MeasureRiseTimeTMU.cs
@@ -40,7 +40,7 @@ namespace NationalInstruments.Examples.SemiconductorTestLibrary.TMU
         public static void MeasureRiseTimeWithSTL(ISemiconductorModuleContext tsmContext)
         {
             // Configuration parameters for TMU rise time measurement.
-            int numberOfSamples = 100;           // Number of rise time samples to collect.
+            long numberOfSamples = 100;           // Number of rise time samples to collect.
             double timeoutInSeconds = 5.0;       // Maximum time to wait for measurement completion.
 
             // Step 1: Query TSM session manager to get the digital sessions bundle associated with the "C0" pin.
